+++ src/pyapollo/__init__.py
@@ -11,15 +11,33 @@ from pyapollo.async_ import AsyncApolloClient
 from pyapollo.config import ApolloSettingsConfig
 from pyapollo.core.models import ChangeType, ConfigChange, ConfigChangeEvent
 from pyapollo.listeners import Subscription
+from pyapollo.provider import ApolloConfigProvider
+from pyapollo.registry import (
+    get_client,
+    init_apollo,
+    is_apollo_initialized,
+    register_apollo_factory,
+    require_client,
+    reset_apollo,
+    shutdown_apollo,
+)
 
 logging.getLogger("pyapollo").addHandler(logging.NullHandler())
 
 __all__ = [
     "ApolloClient",
+    "ApolloConfigProvider",
     "ApolloSettingsConfig",
     "AsyncApolloClient",
     "ChangeType",
     "ConfigChange",
     "ConfigChangeEvent",
     "Subscription",
+    "get_client",
+    "init_apollo",
+    "is_apollo_initialized",
+    "register_apollo_factory",
+    "require_client",
+    "reset_apollo",
+    "shutdown_apollo",
 ]